+++ src/Wait.php
@@ -12,6 +12,8 @@ use Innmind\TimeContinuum\{
 };
 use Innmind\Stream\{
     Stream,
+    Readable,
+    Writable,
     Watch\Ready,
 };
 use Innmind\Immutable\{
@@ -33,21 +35,41 @@ final class Wait
     private Maybe $source;
     /** @var Sequence<Task\PendingActivity<R>> */
     private Sequence $tasks;
+    /** @var Maybe<ElapsedPeriod> */
+    private Maybe $timeout;
+    /** @var Set<Readable> */
+    private Set $forRead;
+    /** @var Set<Writable> */
+    private Set $forWrite;
+    /** @var Set<Readable> */
+    private Set $files;
     private bool $poll;
 
     /**
      * @param Maybe<Task\PendingActivity<Context<C, R>>> $source
      * @param Sequence<Task\PendingActivity<R>> $tasks
+     * @param Maybe<ElapsedPeriod> $timeout
+     * @param Set<Readable> $forRead
+     * @param Set<Writable> $forWrite
+     * @param Set<Readable> $files
      */
     private function __construct(
         OperatingSystem $os,
         Maybe $source,
         Sequence $tasks,
+        Maybe $timeout,
+        Set $forRead,
+        Set $forWrite,
+        Set $files,
         bool $poll,
     ) {
         $this->os = $os;
         $this->source = $source;
         $this->tasks = $tasks;
+        $this->timeout = $timeout;
+        $this->forRead = $forRead;
+        $this->forWrite = $forWrite;
+        $this->files = $files;
         $this->poll = $poll;
     }
 
@@ -62,34 +84,10 @@ final class Wait
         $started = $this->os->clock()->now();
         $shortestTimeout = match ($this->poll) {
             true => Maybe::just(Earth\ElapsedPeriod::of(0)),
-            false => $this->shortestTimeout(),
+            false => $this->timeout,
         };
-        $forRead = $this
-            ->tasks
-            ->map(static fn($task) => $task->action()->forRead())
-            ->toSet()
-            ->flatMap(static fn($streams) => $streams)
-            ->merge(
-                $this
-                    ->source
-                    ->toSequence()
-                    ->toSet()
-                    ->flatMap(static fn($task) => $task->action()->forRead()),
-            );
-        $forWrite = $this
-            ->tasks
-            ->map(static fn($task) => $task->action()->forWrite())
-            ->toSet()
-            ->flatMap(static fn($streams) => $streams)
-            ->merge(
-                $this
-                    ->source
-                    ->toSequence()
-                    ->toSet()
-                    ->flatMap(static fn($task) => $task->action()->forWrite()),
-            );
 
-        if ($forRead->empty() && $forWrite->empty()) {
+        if ($this->forRead->empty() && $this->forWrite->empty()) {
             $_ = $shortestTimeout->match(
                 fn($shortestTimeout) => $this
                     ->os
@@ -109,13 +107,12 @@ final class Wait
             static fn($timeout) => $timeout,
             static fn() => null,
         );
-        /** @psalm-suppress InvalidArgument */
         $watch = $this->os->sockets()->watch($timeout);
-        $watch = $forRead->match(
+        $watch = $this->forRead->match(
             static fn($read, $rest) => $watch->forRead($read, ...$rest->toList()),
             static fn() => $watch,
         );
-        $watch = $forWrite->match(
+        $watch = $this->forWrite->match(
             static fn($write, $rest) => $watch->forWrite($write, ...$rest->toList()),
             static fn() => $watch,
         );
@@ -133,8 +130,8 @@ final class Wait
         // be flagged as EOF.
         // This hack SHOULD NOT pose any side effects, however if you land here
         // it may mean it does. In such case I have no idea how to fix it :/
-        $ready = $watch()->map(static fn($ready) => new Ready(
-            $ready->toRead()->merge($forRead->filter(self::isFile(...))),
+        $ready = $watch()->map(fn($ready) => new Ready(
+            $ready->toRead()->merge($this->files),
             $ready->toWrite(),
         ));
         $took = $this->os->clock()->now()->elapsedSince($started);
@@ -149,8 +146,19 @@ final class Wait
     {
         /** @var Maybe<Task\PendingActivity<Context<mixed, mixed>>> */
         $source = Maybe::nothing();
+        /** @var Maybe<ElapsedPeriod> */
+        $timeout = Maybe::nothing();
 
-        return new self($os, $source, Sequence::of(), false);
+        return new self(
+            $os,
+            $source,
+            Sequence::of(),
+            $timeout,
+            Set::of(),
+            Set::of(),
+            Set::of(),
+            false,
+        );
     }
 
     /**
@@ -163,14 +171,27 @@ final class Wait
      */
     public function withSource(Maybe $source): self
     {
+        $sourceTask = $source->keep(Instance::of(Task\PendingActivity::class));
+        $forRead = $sourceTask->match(
+            static fn($task) => $task->action()->forRead(),
+            static fn() => Set::of(),
+        );
+
         /**
          * @psalm-suppress MixedArgumentTypeCoercion Force erase the type on purpose
          * @var self<A, B>
          */
         return new self(
             $this->os,
-            $source->keep(Instance::of(Task\PendingActivity::class)),
+            $sourceTask,
             $this->tasks,
+            $sourceTask->flatMap(static fn($source) => $source->action()->timeout()),
+            $forRead,
+            $sourceTask->match(
+                static fn($task) => $task->action()->forWrite(),
+                static fn() => Set::of(),
+            ),
+            $forRead->filter(self::isFile(...)),
             $source
                 ->keep(Instance::of(Context::class))
                 ->match(
@@ -187,10 +208,26 @@ final class Wait
      */
     public function with(Task\PendingActivity $task): self
     {
+        $timeout = $task->action()->timeout();
+
         return new self(
             $this->os,
             $this->source,
             $this->tasks->add($task),
+            $this
+                ->timeout
+                ->flatMap(static fn($shortest) => $timeout->map(
+                    static fn($new) => match ($shortest->longerThan($new)) {
+                        true => $new,
+                        false => $shortest,
+                    },
+                ))
+                ->otherwise(static fn() => $timeout),
+            $this->forRead->merge($task->action()->forRead()),
+            $this->forWrite->merge($task->action()->forWrite()),
+            $this->files->merge(
+                $task->action()->forRead()->filter(self::isFile(...)),
+            ),
             $this->poll,
         );
     }
@@ -217,31 +254,6 @@ final class Wait
         ));
 
         return [$source, $tasks];
-    }
-
-    /**
-     * @return Maybe<ElapsedPeriod>
-     */
-    private function shortestTimeout(): Maybe
-    {
-        $shortestTimeout = $this
-            ->source
-            ->flatMap(static fn($source) => $source->action()->timeout());
-
-        return $this
-            ->tasks
-            ->map(static fn($task) => $task->action()->timeout())
-            ->reduce(
-                $shortestTimeout,
-                static fn(Maybe $shortestTimeout, Maybe $timeout) => $shortestTimeout
-                    ->flatMap(static fn(ElapsedPeriod $a) => $timeout->map(
-                        static fn(ElapsedPeriod $b) => match ($a->longerThan($b)) {
-                            true => $b,
-                            false => $a,
-                        },
-                    ))
-                    ->otherwise(static fn() => $timeout),
-            );
     }
 
     private static function isFile(Stream $stream): bool